+++ audit/lambda/s3Services/getS3ObjectAsString.ts
