+++ audit/lambda/s3Services/putS3Object.ts
